implement dropping
Dropping required changing the auxiliar script for it to handle both
right-click menu and the different kinds of dropping (shift+drop for
moving, ctrl+drop for coppying, etc).

The auxiliar script is now called `xfilesctl` and has the following
invocation syntax:

	xfilesctl COMMAND FILES ...

- If COMMAND is "menu", it pops up the right-click menu.
- If COMMAND is "drop-ask", it pops up the drop menu.
- If COMMAND is "drop-copy", it copies the dropped files.
Etc.

The auxiliar script will be augmented if further interactions and
controls are implemented in the future.